+++ Win32Game/AttackTable.h
@@ -27,7 +27,7 @@ struct AttackInfo {
   bool isMultiHit_ = false;
   float damage_ = 0.0f;
   Vector knockBackForce_ = {0.0f, 0.0f};
-  float freezeTime_ = 0.0f;
+  unsigned long long freezeTime_ = 0;
   void* searchHandle_ = nullptr;
 };
 
@@ -43,7 +43,15 @@ class AttackTable
 
   bool Initailize();
 
-  bool RegistAttackInfo(unsigned long long attackTableKey, ATTACK_TYPE attackType, ATTACK_ELEMENT_TYPE elementType, EFFECT_TYPE effectType, bool isMultiHit, float damage, Vector knockBackForce, float freezeTime);
+  bool RegistAttackInfo(
+      unsigned long long attackTableKey, 
+      ATTACK_TYPE attackType, 
+      ATTACK_ELEMENT_TYPE elementType, 
+      EFFECT_TYPE effectType, 
+      bool isMultiHit, 
+      float damage, 
+      Vector knockBackForce, 
+      unsigned long long freezeTime);
 
   bool SearchAttackInfo(unsigned long long attackTableKey, AttackInfo** outSearchResult);
